Major Level Changes, Fixes
Made back up files of my first pass CSVs for each level, then modified level CSVs to stop spawning enemies from bottom of the screen to gauge difference in gameplay experience.

Altered badgerBoss and phantomBoss prefabs to limit floaty movement to a smaller portion of the screen because player cannot shoot them when they hang out at the bottom of the screen.

Moved player spawner to the center of the playable area of the screen.